+++ cloud/acp-spring-cloud-starter/src/main/kotlin/pers/acp/spring/cloud/AcpCloudComponentAutoConfiguration.kt
@@ -1,8 +1,10 @@
 package pers.acp.spring.cloud
 
 import com.fasterxml.jackson.databind.ObjectMapper
+import com.netflix.hystrix.strategy.concurrency.HystrixConcurrencyStrategy
 import feign.RequestInterceptor
 import org.springframework.beans.factory.config.ConfigurableBeanFactory
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.context.annotation.*
 import org.springframework.http.HttpHeaders
 import org.springframework.web.context.request.RequestContextHolder
@@ -43,6 +45,7 @@ class AcpCloudComponentAutoConfiguration {
     }
 
     @Bean
+    @ConditionalOnMissingBean(HystrixConcurrencyStrategy::class)
     fun feignHystrixConcurrencyStrategy() = FeignHystrixConcurrencyStrategy()
 
     @Bean