+++ cloud/acp-spring-cloud-starter/src/main/kotlin/pers/acp/spring/cloud/AcpCloudLogServerClientAutoConfiguration.kt
@@ -1,10 +1,8 @@
 package pers.acp.spring.cloud
 
-import com.fasterxml.jackson.databind.ObjectMapper
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.autoconfigure.AutoConfigureBefore
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.cloud.stream.annotation.EnableBinding
 import org.springframework.cloud.stream.config.BindingProperties
 import org.springframework.cloud.stream.config.BindingServiceConfiguration
@@ -12,11 +10,8 @@ import org.springframework.cloud.stream.config.BindingServiceProperties
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.util.MimeTypeUtils
-import pers.acp.spring.boot.AcpBootLogAutoConfiguration
-import pers.acp.spring.boot.interfaces.LogAdapter
 import pers.acp.spring.cloud.conf.AcpCloudLogServerClientConfiguration
 import pers.acp.spring.cloud.conf.AcoCloudLogServerConfiguration
-import pers.acp.spring.cloud.log.CloudLogAdapter
 import pers.acp.spring.cloud.log.producer.LogOutput
 import pers.acp.spring.cloud.log.producer.LogProducer
 import pers.acp.spring.cloud.log.LogConstant
@@ -31,10 +26,12 @@ import javax.annotation.PostConstruct
  */
 @Configuration
 @ConditionalOnExpression("'\${acp.cloud.log-server.client.enabled}'.equals('true')")
-@AutoConfigureBefore(BindingServiceConfiguration::class, AcpBootLogAutoConfiguration::class)
+@AutoConfigureBefore(BindingServiceConfiguration::class)
 @EnableBinding(LogOutput::class)
 class AcpCloudLogServerClientAutoConfiguration @Autowired
-constructor(private val acoCloudLogServerConfiguration: AcoCloudLogServerConfiguration, private val acpCloudLogServerClientConfiguration: AcpCloudLogServerClientConfiguration, private val bindings: BindingServiceProperties) {
+constructor(private val acoCloudLogServerConfiguration: AcoCloudLogServerConfiguration,
+            private val acpCloudLogServerClientConfiguration: AcpCloudLogServerClientConfiguration,
+            private val bindings: BindingServiceProperties) {
 
     /**
      * 初始化日志消息生产者
@@ -56,10 +53,5 @@ constructor(private val acoCloudLogServerConfiguration: AcoCloudLogServerConfigu
 
     @Bean
     fun logProducer(logOutput: LogOutput): LogProducer = LogProducer(logOutput)
-
-    @Bean
-    @ConditionalOnMissingBean(LogAdapter::class)
-    fun logAdapter(acpCloudLogServerClientConfiguration: AcpCloudLogServerClientConfiguration, objectMapper: ObjectMapper) =
-            CloudLogAdapter(acpCloudLogServerClientConfiguration, objectMapper)
 
 }