+++ doc/version_history.md
@@ -4,6 +4,7 @@
 > - [Upgrade] 优化部分代码
 > - [Upgrade] 调整 acp-spring-boot-starter 及 acp-spring-cloud-starter 自动配置写法，去除 ComponentScan 包扫描
 > - [Upgrade] acp-spring-boot-starter 中增加日志适配器接口，根据不同场景（独立SpringBoot项目SpringCloud项目）进行更换日志输出方式，并在 acp-spring-boot-starter 及 acp-spring-cloud-starter 中分别实现了日志适配器
+> - [Upgrade] 优化自动配置写法，方便扩展
 ##### v6.0.1
 > - [Upgrade] 优化 RestControllerAspect 切面日志处理
 > - [Upgrade] testspringboot 增加异步请求接口demo